third_party/WebKit/LayoutTests/platform/win/http/tests/htmlimports/redirect-cross-origin-cross-same-2-expected.txt

CONSOLE ERROR: Access to Imported resource at 'http://127.0.0.1:8000/htmlimports/resources/hello.html' from origin 'null' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'null' is therefore not allowed access.
This is a testharness.js-based test.
PASS Accessing to a same origin import including intermediate CORS domain redirect 
Harness: the test ran to completion.
